Align timeline settings utility conventions
