+++ packages/twenty-front/src/pages/settings/applications/types/settingsApplicationTimelineActivityType.ts
@@ -1,4 +1,3 @@
-import { type TimelineActivityTypeManifest } from 'twenty-shared/application';
 import { type FindManyTimelineActivityTypesQuery } from '~/generated-metadata/graphql';
 
 export type InstalledTimelineActivityType =
@@ -13,39 +12,3 @@ export type SettingsApplicationTimelineActivityType = {
   label: string;
   name: string;
 };
-
-export const getSettingsApplicationTimelineActivityTypes = ({
-  applicationId,
-  installedApplication,
-  installedTimelineActivityTypes,
-  manifestTimelineActivityTypes,
-}: {
-  applicationId: string;
-  installedApplication: boolean;
-  installedTimelineActivityTypes: InstalledTimelineActivityType[];
-  manifestTimelineActivityTypes: TimelineActivityTypeManifest[];
-}): SettingsApplicationTimelineActivityType[] =>
-  installedApplication
-    ? installedTimelineActivityTypes
-        .filter(
-          (timelineActivityType) =>
-            timelineActivityType.applicationId === applicationId,
-        )
-        .map((timelineActivityType) => ({
-          action: timelineActivityType.emit?.on ?? null,
-          icon: timelineActivityType.icon,
-          id: timelineActivityType.id,
-          isActive: timelineActivityType.isActive,
-          isInstalled: true,
-          label: timelineActivityType.label,
-          name: timelineActivityType.name,
-        }))
-    : manifestTimelineActivityTypes.map((timelineActivityType) => ({
-        action: timelineActivityType.emit?.on ?? null,
-        icon: timelineActivityType.icon,
-        id: timelineActivityType.universalIdentifier,
-        isActive: true,
-        isInstalled: false,
-        label: timelineActivityType.label,
-        name: timelineActivityType.name,
-      }));
